✅ TestContainerConfig 정리
