+++ build.gradle.kts
@@ -38,8 +38,6 @@ dependencies {
 	compileOnly("org.projectlombok:lombok")
 	annotationProcessor("org.projectlombok:lombok")
 
-	implementation("org.springdoc:springdoc-openapi-starter-webmvc-ui:2.0.2")
-
 	// swagger
 	implementation("org.springdoc:springdoc-openapi-starter-webmvc-ui:2.7.0")
 
@@ -52,15 +50,14 @@ dependencies {
 	// Test
 	testImplementation("org.springframework.boot:spring-boot-starter-test")
 	testImplementation("org.springframework.boot:spring-boot-testcontainers")
-	testImplementation("com.redis.testcontainers:testcontainers-redis:1.6.4")
 	testImplementation("org.mockito:mockito-inline:5.2.0")
 	testImplementation("org.testcontainers:junit-jupiter")
 	testImplementation("org.testcontainers:mysql")
 	testImplementation("org.testcontainers:kafka")
+	testImplementation("com.redis:testcontainers-redis")
 	testRuntimeOnly("org.junit.platform:junit-platform-launcher")
 }
 
 tasks.withType<Test> {
 	useJUnitPlatform()
-	systemProperty("user.timezone", "UTC")
 }
